Movie-rating-System
MovieHub aims to provide a user-friendly and interactive platform for movie enthusiasts to discover, rate, and discuss their favorite films. The project showcases the capabilities of HTML and CSS in creating a visually appealing and responsive front end, while the use of a robust back-end stack ensures the security and functionality of the overall system. MovieHub is not just a movie rating system but a community hub for cinephiles to connect and share their passion for cinema.